Improve test coverage in test_hl_line()

+++ test/data/test.c
@@ -2,6 +2,6 @@
 
 int c_syntax_test(int argc, char *argv[])
 {
-    fprintf(stderr, "test: %s %c %d\n", "str", 'x', 42);
+    fprintf(stderr, "test: %s %c %zu\n", "str", 'x', (size_t)42);
     return 0;
 }
